refactor: move MinGreaterThanZero function to utils package
